python scripts now using GEM_PYTHON_PATH, which is set in paths.sh
