+++ gemdaq-testing/setup/etc/addresstables/counter_registers.xml
@@ -42,7 +42,7 @@
 		     0x40030X63"/> 
   
   <node id="RESETS"  address="0xC"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/counter_reset_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/counter_reset_registers.xml"
 	description="resets for all the counters"/>
   
 </node> <!-- end trigger setup definitions -->
