+++ gemdaq-testing/setup/etc/addresstables/glib_address_table.xml
@@ -3,7 +3,7 @@
 <node id="top">
   <node id="GLIB">
     <node id="SYSTEM"  address="0x00000000"  description="GLIB system registers"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_system_registers.xml">
+	  module="file://${GEM_ADDRESS_TABLE_PATH/glib_system_registers.xml">
     </node> <!--end system registers -->
 
     <node id="SRAM1"  address="0x02000000"  mask="0xffffffff"  permission="rw"
@@ -17,37 +17,37 @@
 	  description="ICAP memory space (seems to not work at the moment)"/>
     
     <node id="FLASH"  address="0x08000000"  description="FLASH memory banks"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_flash_bank_registers.xml"/>
+	  module="file://${GEM_ADDRESS_TABLE_PATH/glib_flash_bank_registers.xml"/>
 
     <!-- registers for the VFAT tracking data -->
     <node id="TRK_DATA"  address="0x40020000"
 	  description="VFAT tracking data registers">
       
       <node id="COL0"  address="0x000"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/vfat_tracking_data.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/vfat_tracking_data.xml"
 	    description="First phi column">
       </node> <!-- end column 0 block -->
       <node id="COL1"  address="0x100"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/vfat_tracking_data.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/vfat_tracking_data.xml"
 	    description="First phi column">
       </node> <!-- end column 0 block -->
       <node id="COL2"  address="0x200"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/vfat_tracking_data.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/vfat_tracking_data.xml"
 	    description="First phi column">
       </node> <!-- end column 0 block -->
     </node> <!-- end tracking_data block -->
     
     <node id="GLIB_LINKS"  address="0x40040000">
       <node id="LINK0"  address="0x000"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/glib_optical_links.xml"
 	    description="optical link status counters and resets"/>
       
       <node id="LINK1"  address="0x100"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/glib_optical_links.xml"
 	    description="optical link status counters and resets"/>
       
       <node id="LINK2"  address="0x200"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/glib_optical_links.xml"
 	    description="optical link status counters and resets"/>
       
       <node id="TRG_DATA"
