+++ gemdaq-testing/setup/etc/addresstables/glib_system_registers.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="ISO-8859-1"?>
 <node id="top">
   <node id="BOARD_ID" address="0x0"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/four_char_id.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/four_char_id.xml"
 	description="GLIB board ID code">
   </node> <!--end board_id -->
   
   <node id="SYSTEM_ID"  address="0x1"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/four_char_id.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/four_char_id.xml"
 	description="GLIB system ID code">
   </node> <!--end system_id -->
   
@@ -24,7 +24,7 @@
 	description="Register for testing purposes only"/>
 
   <node id="CLK_CTRL"  address="0x4"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_clk_ctrl.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/glib_clk_ctrl.xml"
 	description="Controls the external clocking circuitry">
   </node> <!--end ctrl-->
 
@@ -35,22 +35,22 @@
   </node> <!--end ctrl_2-->
   
   <node id="STATUS"  address="0x6"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_status_register.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/glib_status_register.xml"
 	description="Status from various external components"/>
 
   <node id="STATUS_2"   address="0x7"  mask="0xffffffff"  permission="r"
 	description="Currently unused"/>
 
   <node id="SRAM" address="0x8"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_sram_ctrl_status_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/glib_sram_ctrl_status_registers.xml"
 	description="SRAM interface control and status registers, and banks 1 and 2"/>
   
   <node id="SPI"  address="0xa"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_spi_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/glib_spi_registers.xml"
 	description="SPI interface: data from FPGA to clock synthesizer"/>
   
   <node id="I2C"  address="0xd"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_i2c_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/glib_i2c_registers.xml"
 	description="I2C interface"/>
 
   <node id="SFP_PHASE_MON">
