+++ gemdaq-testing/setup/etc/addresstables/optohybrid_address_table.xml
@@ -7,37 +7,37 @@
 	  description="Optohybrid registers">
 
       <node id="LINK0"  address="0x000"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/optohybrid_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/optohybrid_optical_links.xml"
 	    description="optical link status counters and resets"/>
       
       <node id="LINK1"  address="0x100"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/optohybrid_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/optohybrid_optical_links.xml"
 	    description="optical link status counters and resets"/>
       
       <node id="LINK2"  address="0x200"
-	    module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/optohybrid_optical_links.xml"
+	    module="file://${GEM_ADDRESS_TABLE_PATH/optohybrid_optical_links.xml"
 	    description="optical link status counters and resets"/>
     </node> <!-- end OptoHybrid links block-->
     
     <node id="GEB"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/geb_vfat_address_table.xml"
+	  module="file://${GEM_ADDRESS_TABLE_PATH/geb_vfat_address_table.xml"
 	  description="complete GEB address table"/>
     
     <!--
     <node id="FAST_COM"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/fast_signals_address_table.xml"
+	  module="file://${GEM_ADDRESS_TABLE_PATH/fast_signals_address_table.xml"
 	  description="global fast commands registers"/>
     
     <node id="TRIGGER"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/trigger_registers.xml"
+	  module="file://${GEM_ADDRESS_TABLE_PATH/trigger_registers.xml"
 	  description="global ttrigger setup"/>
     
     <node id="CLOCKING"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/clocking_registers.xml"
+	  module="file://${GEM_ADDRESS_TABLE_PATH/clocking_registers.xml"
 	  description="global clocking registers"/>
     
     <node id="COUNTERS"
-	  module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/counter_registers.xml"
+	  module="file://${GEM_ADDRESS_TABLE_PATH/counter_registers.xml"
 	  description="global counter registers"/>
     
     <node id="FIRMWARE"  address="0x40030157"  mask="0xFFFFFFFF"  permission="r"
