+++ gemdaq-testing/setup/etc/addresstables/optohybrid_optical_links.xml
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="ISO-8859-1"?>
 <node id="top">
   <node id="OPTICAL_LINKS"  address="0x00"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/optical_link_registers.xml"/>
+	module="file://${GEM_ADDRESS_TABLE_PATH/optical_link_registers.xml"/>
   
   <node id="FAST_COM"  address="0x40"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/fast_signals_address_table.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/fast_signals_address_table.xml"
 	description="global fast commands registers"/>
   
   <node id="TRIGGER"   address="0x45"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/trigger_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/trigger_registers.xml"
 	description="global trigger setup"/>
   
   <node id="CLOCKING"  address="0x4B"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/clocking_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/clocking_registers.xml"
 	description="global clocking registers"/>
   
   <node id="COUNTERS"  address="0x5A"
-	module="file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/counter_registers.xml"
+	module="file://${GEM_ADDRESS_TABLE_PATH/counter_registers.xml"
 	description="global counter registers"/>
   
   <node id="FIRMWARE"  address="0x57"  mask="0xFFFFFFFF"  permission="r"
