+++ gemdaq-testing/setup/paths.sh
@@ -1,4 +1,7 @@
 #export BUILD_HOME=</path/to/your/BUILD_HOME>
+if [[ -n "$BUILD_HOME" ]]; then
+    echo "BUILD_HOME not set, please set BUILD_HOME to the root of your repository and rerun this script"
+    exit(1)
 #Run this script after you've set BUILD_HOME, or uncomment the previous line and set it to your BUILD_HOME
 
 # The OS is not set in environment. We assume we are not cross-compiling, and try
@@ -48,3 +51,6 @@ export LD_LIBRARY_PATH=$BUILD_HOME/$project/gemhardware/lib/$XDAQ_OS/$XDAQ_PLATF
 export LD_LIBRARY_PATH=$BUILD_HOME/$project/gemsupervisor/lib/$XDAQ_OS/$XDAQ_PLATFORM:$LD_LIBRARY_PATH
 export LD_LIBRARY_PATH=$BUILD_HOME/$project/gemHwMonitor/lib/$XDAQ_OS/$XDAQ_PLATFORM:$LD_LIBRARY_PATH
 echo LD_LIBRARY_PATH $LD_LIBRARY_PATH
+export GEM_ADDRESS_TABLE_PATH=${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables
+export GEM_PYTHON_PATH=${BUILD_HOME}/gemdaq-testing/setup/scripts/python
+export PYTHONPATH=${GEM_PYTHON_PATH}:${PYTHONPATH}
