+++ gemdaq-testing/setup/scripts/python/optohybrid_board_info_uhal.py
@@ -3,7 +3,7 @@
 import sys, re
 import time, datetime, os
 
-sys.path.append('${BUILD_HOME}/gemdaq-testing/setup/scripts')
+sys.path.append('${GEM_PYTHON_PATH')
 
 import uhal
 from registers_uhal import *
@@ -44,10 +44,10 @@ if options.slot:
 	uTCAslot = 160+options.slot
 	print options.slot, uTCAslot
 ipaddr = '192.168.0.%d'%(uTCAslot)
-address_table = "file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/optohybrid_address_table.xml"
+address_table = "file://${GEM_ADDRESS_TABLE_PATH/optohybrid_address_table.xml"
 uri = "chtcp-2.0://localhost:10203?target=%s:50001"%(ipaddr)
 optohybrid  = uhal.getDevice( "optohybrid" , uri, address_table )
-address_table = "file://${BUILD_HOME}/gemdaq-testing/setup/etc/addresstables/glib_address_table.xml"
+address_table = "file://${GEM_ADDRESS_TABLE_PATH/glib_address_table.xml"
 glib  = uhal.getDevice( "glib" , uri, address_table )
 
 ########################################
