+++ gemdaq-testing/setup/scripts/python/registers_uhal.py
@@ -1,7 +1,7 @@
 import sys, os, time, signal, random
-sys.path.append('${BUILD_HOME}/gemdaq-testing/setup/scripts')
+sys.path.append('${GEM_PYTHON_PATH')
 
-from PyChipsUser import *
+import uhal
 
 gMAX_RETRIES = 11
 gRetries = 0
@@ -16,38 +16,56 @@ class colors:
 	ENDC    = '\033[0m'
 
 
-def readRegister(glib,register):
+def readRegister(device,register):
+	"""
+	read register 'register' from uhal device 'device'
+	returns value of the register
+	"""
 	global gRetries
 	nRetries = 0
 	while (nRetries < gMAX_RETRIES):
 		try:
-			controlChar = glib.read(register)
+			controlChar = device.getNode(register).read()
+			device.dispatch()
 			return controlChar
-		
-		except ChipsException, e:
+	#want to be able to return nothing in the result of a failed transaction
+		except uhal.exception, e:
 			nRetries += 1
 			gRetries += 1
+			#if ('amount of data' in e):
+			#	print colors.BLUE, "bad header",register, "-> Error : ", e, colors.ENDC
+			#elif ('INFO CODE = 0x4L' in e):
+			#	print colors.CYAN, "read error",register, "-> Error : ", e, colors.ENDC
+			#elif ('INFO CODE = 0x6L' in e or 'timed out' in e):
+			#	print colors.YELLOW, "timed out",register, "-> Error : ", e, colors.ENDC
+			#else:
+			#	print colors.MAGENTA, "other error",register, "-> Error : ", e, colors.ENDC
 			if ((nRetries % 10)==0):
 				print colors.MAGENTA,"read error encountered (%s), retrying operation (%d,%d)"%(register,nRetries,gRetries),e,colors.ENDC
 			continue
 		pass
 	#print colors.RED, "error encountered, retried read operation (%d)"%(nRetries)
 	return 0x0
+	# return 0x0
 	
-def writeRegister(glib,register, value):
+def writeRegister(device,register, value):
+	"""
+	write value 'value' into register 'register' from uhal device 'device'
+	"""
 	global gRetries
 	nRetries = 0
 	while (nRetries < gMAX_RETRIES):
 		try:
-			glib.write(register, value)
+			device.getNode(register).write(value)
+			device.dispatch()
 			return
-
-		except ChipsException, e:
-			#if ('amount of data' in e.value):
+		
+		except uhal.exception, e:
+			#if ('amount of data' in e):
 			#	print colors.BLUE, "bad header",register, "-> Error : ", e, colors.ENDC
-			#elif ('INFO CODE = 0x4L' in e.value):
+			#elif ('INFO CODE = 0x4L' in e):
 			#	print colors.CYAN, "read error",register, "-> Error : ", e, colors.ENDC
-			#elif ('INFO CODE = 0x6L' in e.value or 'timed out' in e.value):
+			#elif ('INFO CODE = 0x6L' in e or 'timed out' in e):
 			#	print colors.YELLOW, "timed out",register, "-> Error : ", e, colors.ENDC
 			#else:
 			#	print colors.MAGENTA, "other error",register, "-> Error : ", e, colors.ENDC
@@ -59,4 +77,3 @@ def writeRegister(glib,register, value):
 		pass
 	#print colors.RED, "error encountered, retried test write operation (%d)"%(nRetries)
 	pass
-
